CP2000_Pipeline/requirements.txt
PyMuPDF==1.23.5
openpyxl==3.1.2
pandas==2.0.3
requests==2.31.0
google-api-python-client==2.100.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
python-dateutil==2.8.2
watchdog==3.0.0